Kenzie/Sub-MAF-07: Fixed on start bugs
