+++ app/src/main/java/com.terracode.blueharvest/ConfigurationSettingsActivity.kt
@@ -1,29 +1,46 @@
 package com.terracode.blueharvest
 
 import android.content.Intent
+import android.content.SharedPreferences
 import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
 import android.view.View
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
+import androidx.preference.PreferenceManager
+import com.terracode.blueharvest.utils.LocaleHelper
 import com.terracode.blueharvest.utils.SetTextSize
 import com.terracode.blueharvest.utils.ThemeHelper
 
 class ConfigurationSettingsActivity : AppCompatActivity() {
+    // Declare variables as var to allow reassignment
+    private lateinit var sharedPreferences: SharedPreferences
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
+        //Initialize the sharedPreferences
+        sharedPreferences = PreferenceManager.getDefaultSharedPreferences(this)
+
+        //Set setting values before setting the content view
         val currentTheme = ThemeHelper.getCurrentTheme(this)
         ThemeHelper.setColorOverlayTheme(this, currentTheme)
+
+        val currentLanguagePosition = sharedPreferences.getInt("selectedLanguagePosition", 0)
+        val languagePosition = LocaleHelper.getLanguageCode(currentLanguagePosition)
+        LocaleHelper.setLocale(this, languagePosition)
+
+        //Set the view
         setContentView(R.layout.activity_configuration_settings)
 
+        //Set the toolbar
         val toolbar: Toolbar = findViewById(R.id.settingsToolbar)
         setSupportActionBar(toolbar)
 
+        //Set the text size
         val rootView = findViewById<View>(android.R.id.content).rootView
         SetTextSize.setTextSizeView(this, rootView)
-
     }
 
     override fun onCreateOptionsMenu(menu: Menu): Boolean {
