+++ app/src/main/java/com.terracode.blueharvest/listeners/LanguageListener.kt
@@ -1,13 +1,12 @@
 package com.terracode.blueharvest.listeners
 
-import android.content.res.Configuration
 import android.view.View
 import android.widget.AdapterView
 import androidx.preference.PreferenceManager
 import com.terracode.blueharvest.AccessibilitySettingsActivity
-import java.util.Locale
+import com.terracode.blueharvest.utils.LocaleHelper
 
-@Suppress("DEPRECATION")
+
 class LanguageSelectionListener(private val activity: AccessibilitySettingsActivity) :
     AdapterView.OnItemSelectedListener {
 
@@ -19,9 +18,9 @@ class LanguageSelectionListener(private val activity: AccessibilitySettingsActiv
     ) {
         val sharedPreferences = PreferenceManager.getDefaultSharedPreferences(activity) // Access preferences within activity context
         val currentLanguagePosition = sharedPreferences.getInt("selectedLanguagePosition", 0)
-        val selectedLanguageCode = getLanguageCode(position)
+        val selectedLanguageCode = LocaleHelper.getLanguageCode(position)
         if (currentLanguagePosition != position) {
-            setLocale(selectedLanguageCode)
+            LocaleHelper.setLocale(activity,selectedLanguageCode)
             sharedPreferences.edit().putInt("selectedLanguagePosition", position).apply()
             activity.recreate()
         }
@@ -29,27 +28,5 @@ class LanguageSelectionListener(private val activity: AccessibilitySettingsActiv
 
     override fun onNothingSelected(parent: AdapterView<*>) {
         // write code to perform some action
-    }
-
-    fun getLanguageCode(position: Int): String {
-        return when (position) {
-            0 -> "en" // English
-            1 -> "fr" // French
-            2 -> "es" // Spanish
-            else -> "en" // Default to English if position is out of range
-        }
-    }
-
-    fun setLocale(languageCode: String) {
-        val locale = Locale(languageCode)
-        Locale.setDefault(locale)
-
-        val resources = activity.resources
-        val configuration = Configuration(resources.configuration)
-        // Set the new locale configuration
-        configuration.setLocale(locale)
-
-        // Update the base context of the application
-        activity.baseContext.resources.updateConfiguration(configuration, activity.baseContext.resources.displayMetrics)
     }
 }
